+++ src/generic/json_structural_indexer.h
@@ -22,7 +22,7 @@ public:
     // it helps tremendously.
     if (bits == 0)
         return;
-    uint32_t cnt = count_ones(bits);
+    int cnt = static_cast<int>(count_ones(bits));
 
     // Do the first 8 all together
     for (int i=0; i<8; i++) {
@@ -42,7 +42,7 @@ public:
       // branch mispredictions here. 16+ structurals per block means either punctuation ({} [] , :)
       // or the start of a value ("abc" true 123) every four characters.
       if (unlikely(cnt > 16)) {
-        uint32_t i = 16;
+        int i = 16;
         do {
           this->tail[i] = idx + trailing_zeroes(bits);
           bits = clear_lowest_bit(bits);
@@ -77,14 +77,14 @@ private:
 really_inline void json_structural_indexer::next(simd::simd8x64<uint8_t> in, json_block block, size_t idx) {
   uint64_t unescaped = in.lteq(0x1F);
   checker.check_next_input(in);
-  indexer.write(idx-64, prev_structurals); // Output *last* iteration's structurals to the parser
+  indexer.write(uint32_t(idx-64), prev_structurals); // Output *last* iteration's structurals to the parser
   prev_structurals = block.structural_start();
   unescaped_chars_error |= block.non_quote_inside_string(unescaped);
 }
 
 really_inline error_code json_structural_indexer::finish(parser &parser, size_t idx, size_t len, bool streaming) {
   // Write out the final iteration's structurals
-  indexer.write(idx-64, prev_structurals);
+  indexer.write(uint32_t(idx-64), prev_structurals);
 
   error_code error = scanner.finish(streaming);
   if (unlikely(error != SUCCESS)) { return error; }
@@ -93,7 +93,7 @@ really_inline error_code json_structural_indexer::finish(parser &parser, size_t 
     return UNESCAPED_CHARS;
   }
 
-  parser.n_structural_indexes = indexer.tail - parser.structural_indexes.get();
+  parser.n_structural_indexes = uint32_t(indexer.tail - parser.structural_indexes.get());
   /* a valid JSON file cannot have zero structural indexes - we should have
    * found something */
   if (unlikely(parser.n_structural_indexes == 0u)) {
@@ -105,7 +105,7 @@ really_inline error_code json_structural_indexer::finish(parser &parser, size_t 
   if (len != parser.structural_indexes[parser.n_structural_indexes - 1]) {
     /* the string might not be NULL terminated, but we add a virtual NULL
      * ending character. */
-    parser.structural_indexes[parser.n_structural_indexes++] = len;
+    parser.structural_indexes[parser.n_structural_indexes++] = uint32_t(len);
   }
   /* make it safe to dereference one beyond this array */
   parser.structural_indexes[parser.n_structural_indexes] = 0;
